requirements.txt
Django==4.1.4
psycopg2==2.9.5
djangorestframework==3.14.0
PyJWT==2.6.0
django-filter==22.1